Tried out a switch statement

+++ number_test.cc
@@ -3,7 +3,7 @@ using namespace std;
 
 int main() {
     int num1;
-    int num2;
+    float num2;
     cin >> num1;
     cin >> num2;
     cout << num1/num2 << endl;
